little bit of cpp style renamings; todo added;
