+++ slot_machine/TODO.txt
@@ -1,0 +1,7 @@
+	1. ќтказатьс€ от glaux
+		a. Ёто библиотека которую делали раньше MS и включали в Windows SDK, сейчас она уже давно устарела, и ее нету в современных SDK
+	2. ѕосмотреть в сторону мета проектов: gyp, cmake, etc
+		a.  россплатформенна€ сборка
+		b. —ложно следить за двум€ конфигураци€ми (релиз, дебаг) руками, кроме того, в ui vs часто трудно найти те или иные настройки, с мета проектов будет описательный текстовый файл, в котором проще работать
+	3. ѕоискать другую библиотеку дл€ jpeg, либо попробовать собирать ее самосто€тельно, т.к. есть проблемы с разными компил€торами, в частности из-за нее нельз€ использовать shared runtime
+	4. —ледить за стилистикой наименовани€ xxx_yyy; xxxYyy; XxxYyy; etc
