+++ slot_machine/button.hpp
