+++ slot_machine/gettexture.hpp
