+++ slot_machine/main.hpp
