+++ slot_machine/reel.hpp
