+++ slot_machine/slot_machine.vcxproj.filters
@@ -1,0 +1,60 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <ItemGroup>
+    <ClCompile Include="texture.cpp" />
+    <ClCompile Include="main.cpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="button.hpp" />
+    <ClInclude Include="reel.hpp" />
+    <ClInclude Include="slotmachine.hpp" />
+    <ClInclude Include="texture.hpp" />
+    <ClInclude Include="main.hpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <Filter Include="resources">
+      <UniqueIdentifier>{f88d4396-2c08-4ba9-b2d5-d0ee2a51934e}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="resources\textures">
+      <UniqueIdentifier>{1e40f7c3-1bf1-43f5-95c9-54c0ec4ab1d7}</UniqueIdentifier>
+    </Filter>
+  </ItemGroup>
+  <ItemGroup>
+    <Image Include="resources\textures\background.jpg">
+      <Filter>resources\textures</Filter>
+    </Image>
+    <Image Include="resources\textures\background1.jpg">
+      <Filter>resources\textures</Filter>
+    </Image>
+    <Image Include="resources\textures\image1.bmp">
+      <Filter>resources\textures</Filter>
+    </Image>
+    <Image Include="resources\textures\image2.jpg">
+      <Filter>resources\textures</Filter>
+    </Image>
+    <Image Include="resources\textures\image3.tga">
+      <Filter>resources\textures</Filter>
+    </Image>
+    <Image Include="resources\textures\selectDis.bmp">
+      <Filter>resources\textures</Filter>
+    </Image>
+    <Image Include="resources\textures\selectEna.bmp">
+      <Filter>resources\textures</Filter>
+    </Image>
+    <Image Include="resources\textures\startDis.bmp">
+      <Filter>resources\textures</Filter>
+    </Image>
+    <Image Include="resources\textures\startEna.bmp">
+      <Filter>resources\textures</Filter>
+    </Image>
+    <Image Include="resources\textures\thefont.png">
+      <Filter>resources\textures</Filter>
+    </Image>
+    <Image Include="resources\textures\thereel.jpg">
+      <Filter>resources\textures</Filter>
+    </Image>
+    <Image Include="resources\textures\thereel1.jpg">
+      <Filter>resources\textures</Filter>
+    </Image>
+  </ItemGroup>
+</Project>
