+++ slot_machine/slotmachine.hpp
