+++ slot_machine/texture.hpp
